requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Django==4.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djoser==2.1.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
model-bakery==1.4.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
